refactor so packets can be verified in ledger
refactor to setup shred signing stage

missing

cleanup

update

fixed test

update
