+++ core/benches/banking_stage.rs
@@ -15,10 +15,10 @@ use solana_core::genesis_utils::{create_genesis_block, GenesisBlockInfo};
 use solana_core::packet::to_packets_chunked;
 use solana_core::poh_recorder::WorkingBankEntry;
 use solana_core::service::Service;
-use solana_core::test_tx::test_tx;
 use solana_ledger::blocktree::{get_tmp_ledger_path, Blocktree};
 use solana_ledger::blocktree_processor::process_entries;
 use solana_ledger::entry::{next_hash, Entry};
+use solana_perf::test_tx::test_tx;
 use solana_runtime::bank::Bank;
 use solana_sdk::genesis_block::GenesisBlock;
 use solana_sdk::hash::Hash;
