+++ core/src/poh_recorder.rs
@@ -469,8 +469,8 @@ impl PohRecorder {
 mod tests {
     use super::*;
     use crate::genesis_utils::{create_genesis_block, GenesisBlockInfo};
-    use crate::test_tx::test_tx;
     use solana_ledger::blocktree::{get_tmp_ledger_path, Blocktree};
+    use solana_perf::test_tx::test_tx;
     use solana_sdk::clock::DEFAULT_TICKS_PER_SLOT;
     use solana_sdk::hash::hash;
     use std::sync::mpsc::sync_channel;
