+++ core/src/poh_service.rs
@@ -112,9 +112,9 @@ mod tests {
     use crate::genesis_utils::{create_genesis_block, GenesisBlockInfo};
     use crate::poh_recorder::WorkingBank;
     use crate::result::Result;
-    use crate::test_tx::test_tx;
     use solana_ledger::blocktree::{get_tmp_ledger_path, Blocktree};
     use solana_ledger::leader_schedule_cache::LeaderScheduleCache;
+    use solana_perf::test_tx::test_tx;
     use solana_runtime::bank::Bank;
     use solana_sdk::hash::hash;
     use solana_sdk::pubkey::Pubkey;
