+++ ledger/Cargo.toml
@@ -16,6 +16,7 @@ chrono = { version = "0.4.9", features = ["serde"] }
 dir-diff = "0.3.2"
 dlopen = "0.1.8"
 dlopen_derive = "0.1.4"
+sha2 = "0.8.0"
 fs_extra = "1.1.0"
 itertools = "0.8.1"
 lazy_static = "1.4.0"
@@ -33,6 +34,7 @@ solana-measure = { path = "../measure", version = "0.21.0" }
 solana-merkle-tree = { path = "../merkle-tree", version = "0.21.0" }
 solana-metrics = { path = "../metrics", version = "0.21.0" }
 solana-perf = { path = "../perf", version = "0.21.0" }
+solana-ed25519-dalek = "0.2.0"
 solana-rayon-threadlimit = { path = "../rayon-threadlimit", version = "0.21.0" }
 solana-runtime = { path = "../runtime", version = "0.21.0" }
 solana-sdk = { path = "../sdk", version = "0.21.0" }
