+++ perf/src/lib.rs
@@ -1,6 +1,16 @@
 pub mod cuda_runtime;
+pub mod packet;
 pub mod perf_libs;
 pub mod recycler;
+pub mod sigverify;
+pub mod test_tx;
 
 #[macro_use]
 extern crate log;
+
+#[cfg(test)]
+#[macro_use]
+extern crate matches;
+
+#[macro_use]
+extern crate solana_metrics;
