+++ perf/src/packet.rs
@@ -1,10 +1,11 @@
 //! The `packet` module defines data structures and methods to pull data from the network.
-use solana_perf::{
+use crate::{
     cuda_runtime::PinnedVec,
     recycler::{Recycler, Reset},
 };
+use serde::Serialize;
 pub use solana_sdk::packet::{Meta, Packet, PACKET_DATA_SIZE};
-use std::{mem, net::SocketAddr};
+use std::{io, mem, net::SocketAddr};
 
 pub const NUM_PACKETS: usize = 1024 * 8;
 
@@ -72,9 +73,42 @@ impl Packets {
     }
 }
 
+pub fn to_packets_chunked<T: Serialize>(xs: &[T], chunks: usize) -> Vec<Packets> {
+    let mut out = vec![];
+    for x in xs.chunks(chunks) {
+        let mut p = Packets::default();
+        p.packets.resize(x.len(), Packet::default());
+        for (i, o) in x.iter().zip(p.packets.iter_mut()) {
+            let mut wr = io::Cursor::new(&mut o.data[..]);
+            bincode::serialize_into(&mut wr, &i).expect("serialize request");
+            let len = wr.position() as usize;
+            o.meta.size = len;
+        }
+        out.push(p);
+    }
+    out
+}
+
+pub fn to_packets<T: Serialize>(xs: &[T]) -> Vec<Packets> {
+    to_packets_chunked(xs, NUM_PACKETS)
+}
+
+pub fn limited_deserialize<T>(data: &[u8]) -> bincode::Result<T>
+where
+    T: serde::de::DeserializeOwned,
+{
+    bincode::config()
+        .limit(PACKET_DATA_SIZE as u64)
+        .deserialize(data)
+}
+
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use solana_sdk::hash::Hash;
+    use solana_sdk::signature::{Keypair, KeypairUtil};
+    use solana_sdk::system_transaction;
 
     #[test]
     fn test_packets_reset() {
@@ -83,5 +117,24 @@ mod tests {
         assert_eq!(packets.packets.len(), 10);
         packets.reset();
         assert_eq!(packets.packets.len(), 0);
+    }
+
+    #[test]
+    fn test_to_packets() {
+        let keypair = Keypair::new();
+        let hash = Hash::new(&[1; 32]);
+        let tx = system_transaction::transfer(&keypair, &keypair.pubkey(), 1, hash);
+        let rv = to_packets(&vec![tx.clone(); 1]);
+        assert_eq!(rv.len(), 1);
+        assert_eq!(rv[0].packets.len(), 1);
+
+        let rv = to_packets(&vec![tx.clone(); NUM_PACKETS]);
+        assert_eq!(rv.len(), 1);
+        assert_eq!(rv[0].packets.len(), NUM_PACKETS);
+
+        let rv = to_packets(&vec![tx.clone(); NUM_PACKETS + 1]);
+        assert_eq!(rv.len(), 2);
+        assert_eq!(rv[0].packets.len(), NUM_PACKETS);
+        assert_eq!(rv[1].packets.len(), 1);
     }
 }
